+++ frontend/src/views/admin/Webhook.vue
@@ -15,11 +15,13 @@ const { t } = useI18n({
             successTip: 'Success',
             webhookAllowList: 'Webhook Allow List(Enter the address that is allowed to use webhook)',
             save: 'Save',
+            notEnabled: 'Webhook is not enabled',
         },
         zh: {
             successTip: '成功',
             webhookAllowList: 'Webhook 白名单(请输入允许使用webhook 的地址)',
             save: '保存',
+            notEnabled: 'Webhook 未开启',
         }
     }
 });
@@ -33,11 +35,13 @@ class WebhookSettings {
 }
 
 const webhookSettings = ref(new WebhookSettings([]))
+const webhookEnabled = ref(false)
 
 const getSettings = async () => {
     try {
         const res = await api.fetch(`/admin/webhook/settings`)
         Object.assign(webhookSettings.value, res)
+        webhookEnabled.value = true
     } catch (error) {
         message.error((error as Error).message || "error");
     }
@@ -62,7 +66,7 @@ onMounted(async () => {
 
 <template>
     <div class="center">
-        <n-card :bordered="false" embedded style="max-width: 800px; overflow: auto;">
+        <n-card v-if="webhookEnabled" :bordered="false" embedded style="max-width: 800px; overflow: auto;">
             <n-form-item-row :label="t('webhookAllowList')">
                 <n-select v-model:value="webhookSettings.allowList" filterable multiple tag
                     :placeholder="t('webhookAllowList')" />
@@ -71,6 +75,7 @@ onMounted(async () => {
                 {{ t('save') }}
             </n-button>
         </n-card>
+        <n-result v-else status="404" :title="t('notEnabled')" />
     </div>
 </template>