+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## main(v0.7.3)
 
 - feat: worker 增加 `ADDRESS_CHECK_REGEX`, address name 的正则表达式, 只用于检查，符合条件将通过检查
+- fix: UI 修复登录页面 tab 激活图标错位
 
 ## v0.7.2